services/courses/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
python-dotenv==1.0.0
django-environ==0.11.2
boto3==1.29.7
django-storages==1.14.2

